fix(user): add actions to user slice and fix button nesting in signup component
