+++ src/pages/Case.tsx
@@ -1,19 +1,18 @@
 import { TabBarCase } from "../components/containers";
 import { Outlet } from "react-router-dom";
-import { useEffect } from 'react';
-import { useAppDispatch } from '../utils/hooks/useStore';
+import { useEffect } from "react";
+import { useAppDispatch } from "../utils/hooks/useStore";
 import { setInsideCase } from "../store/slices/general";
 
 export const Case: React.FC = () => {
-  
   const dispatch = useAppDispatch();
 
-  useEffect( () => {
-    dispatch(setInsideCase(true))
+  useEffect(() => {
+    dispatch(setInsideCase(true));
     return () => {
-      dispatch(setInsideCase(false))
-    }
-  }, [])
+      dispatch(setInsideCase(false));
+    };
+  }, []);
 
   return (
     <>
